+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fixed /docs redirect for root path scenario
+- Fixed /timeseries logging bug
 
 ## [v1.0.0]